agent/pyproject.toml
[project]
name = "agentic-runtime"
version = "0.1.0"
description = "Agentic Kubernetes Operator - Agent Runtime Server using Google ADK"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "litellm>=1.0.0",
    "google-adk[a2a]>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "python-dotenv>=1.0.0",
    "test-mcp-echo-server>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
